pyproject.toml
[project]
name = "ctbc-repo-daily-reload"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "pyinstaller==6.14.1",
    "pyinstaller-hooks-contrib==2025.5",
    "tqdm==4.66.4",
]
